writer-for-fujian/requirements.txt
# 核心依赖
httpx>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# MCP相关
mcp>=0.1.0

# 可选：API服务器
fastapi>=0.104.0
uvicorn>=0.24.0

# 可选：深度检索
# numpy>=1.24.0
# sentence-transformers>=2.2.0
